+++ nextjs-app/src/app/globals.css
@@ -100,7 +100,7 @@ p {
   overflow: hidden;         
   width: auto;              
   white-space: nowrap;  
-  font-size: 24px;
+  font-size: 20px;
   font-weight: 600;
   border-radius: 8px;
 }
